osd/osd_types: skip encoding newer object_info_t fields if they are unused
This reduces the size of the encoded object_info_t in most cases,
enough to get us under the 255 byte limit for a single inline
xattr in XFS.

Note that we lose the local_mtime, but that this field is only
important for objects in the cache pool.

Note that we drop the unconditional clearing of the digest flags if the
encoding was old; this was not strictly correct anyway.  And now that we
may encode the old way with FLAG_OMAP_DIGEST, we need to leave it set (and
force the value to -1).

Signed-off-by: Sage Weil <sage@redhat.com>