config: added thread safe get_val() and SAFE_OPTION macro
The new get_val() takes a lock and returns a copy of the value, thus it's
safe to modify the corresponding setting at the run time (when threads are
already running).
SAFE_OPTION macro declares an option which is safe to change at the run
time. Such an option should be accessed only via get_val()/set_val()
methods.

Signed-off-by: Alexey Sheplyakov <asheplyakov@mirantis.com>